Move drawer items into drawer nav
