+++ src/navigation/BottomTabNavigator/index.js
@@ -1,16 +1,7 @@
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import About from "components/About";
 import ExploreContainer from "components/Explore/ExploreContainer";
-import Identify from "components/Identify/Identify";
 import Messages from "components/Messages/Messages";
-import NetworkLogging from "components/NetworkLogging";
-import PlaceholderComponent from "components/PlaceholderComponent";
-import ProjectDetails from "components/Projects/ProjectDetails";
-import Projects from "components/Projects/Projects";
-import Search from "components/Search/Search";
-import Settings from "components/Settings/Settings";
 import Mortal from "components/SharedComponents/Mortal";
-import UiLibrary from "components/UiLibrary";
 import { t } from "i18next";
 import {
   hideHeader,
@@ -24,7 +15,7 @@ import useUserMe from "sharedHooks/useUserMe";
 
 import CustomTabBar from "./CustomTabBar";
 
-const Tab = createBottomTabNavigator();
+const Tab = createBottomTabNavigator( );
 
 const OBS_LIST_SCREEN_ID = "ObsList";
 const EXPLORE_SCREEN_ID = "Explore";
@@ -46,6 +37,20 @@ const BottomTabs = ( ) => {
         screenOptions={showHeaderLeft}
       >
         <Tab.Screen
+          name="Explore"
+          component={ExploreContainer}
+          options={{
+            ...hideHeader,
+            meta: {
+              icon: "compass-rose-outline",
+              testID: EXPLORE_SCREEN_ID,
+              accessibilityLabel: t( "Explore" ),
+              accessibilityHint: t( "Navigates-to-explore" ),
+              size: 40
+            }
+          }}
+        />
+        <Tab.Screen
           name="ObservationNavigator"
           component={ObservationsStackNavigator}
           options={{
@@ -61,21 +66,6 @@ const BottomTabs = ( ) => {
           }}
         />
         <Tab.Screen
-          name="Explore"
-          component={ExploreContainer}
-          options={{
-            ...hideHeader,
-            meta: {
-              icon: "compass-rose-outline",
-              testID: EXPLORE_SCREEN_ID,
-              accessibilityLabel: t( "Explore" ),
-              accessibilityHint: t( "Navigates-to-explore" ),
-              size: 40
-            }
-          }}
-        />
-
-        <Tab.Screen
           name="Messages"
           component={Messages}
           options={{
@@ -88,68 +78,6 @@ const BottomTabs = ( ) => {
               size: 32
             }
           }}
-        />
-        <Tab.Screen
-          name="search"
-          component={Search}
-          options={{
-            ...hideHeaderLeft,
-            headerTitle: t( "Search" )
-          }}
-        />
-        <Tab.Screen
-          name="Identify"
-          component={Identify}
-          options={{ headerTitle: t( "Identify" ) }}
-        />
-        <Tab.Screen
-          name="Projects"
-          component={Projects}
-          options={{ headerTitle: t( "Projects" ) }}
-        />
-        <Tab.Screen
-          name="ProjectDetails"
-          component={ProjectDetails}
-          options={{ headerTitle: t( "Project" ) }}
-        />
-        <Tab.Screen
-          name="settings"
-          component={Settings}
-          options={{ ...hideHeaderLeft, headerTitle: t( "Settings" ) }}
-        />
-        <Tab.Screen
-          name="about"
-          component={About}
-          options={{ ...hideHeaderLeft, headerTitle: t( "About-iNaturalist" ) }}
-        />
-        <Tab.Screen
-          name="help"
-          component={PlaceholderComponent}
-          options={hideHeaderLeft}
-        />
-        <Tab.Screen
-          name="network"
-          component={NetworkLogging}
-          options={hideHeaderLeft}
-        />
-        <Tab.Screen
-          name="UI Library"
-          component={UiLibrary}
-          options={{
-            ...hideHeaderLeft
-          }}
-        />
-        <Tab.Screen
-          name="Help"
-          component={PlaceholderComponent}
-        />
-        <Tab.Screen
-          name="Blog"
-          component={PlaceholderComponent}
-        />
-        <Tab.Screen
-          name="Donate"
-          component={PlaceholderComponent}
         />
       </Tab.Navigator>
     </Mortal>
